# SAMZA-1464: Flushing a closed RocksDB store causes SIGSEGVs