--- a/samza-core/src/main/java/org/apache/samza/operators/impl/store/TimeSeriesStoreImpl.java
+++ b/samza-core/src/main/java/org/apache/samza/operators/impl/store/TimeSeriesStoreImpl.java
@@ -164,7 +164,6 @@ public class TimeSeriesStoreImpl<K, V> implements TimeSeriesStore<K, V> {
 
   @Override
   public void close() {
-    kvStore.close();
   }
 
   private void validateRange(long startTimestamp, long endTimestamp) throws IllegalArgumentException {