--- a/samza-kv-rocksdb/src/main/scala/org/apache/samza/storage/kv/RocksDbKeyValueStore.scala
+++ b/samza-kv-rocksdb/src/main/scala/org/apache/samza/storage/kv/RocksDbKeyValueStore.scala
@@ -21,6 +21,7 @@ package org.apache.samza.storage.kv
 
 import java.io.File
 import java.util.concurrent.TimeUnit
+import java.util.concurrent.locks.ReentrantReadWriteLock
 
 import org.apache.samza.SamzaException
 import org.apache.samza.config.Config
@@ -103,7 +104,15 @@ class RocksDbKeyValueStore(
   private lazy val db = RocksDbKeyValueStore.openDB(dir, options, storeConfig, isLoggedStore, storeName, metrics)
   private val lexicographic = new LexicographicComparator()
 
-  def get(key: Array[Byte]): Array[Byte] = {
+  /**
+    * null while the store is open. Set to an Exception holding the stacktrace at the time of first close by #close.
+    * Reads and writes to this field must be guarded by stateChangeLock.
+    * This is an Exception instead of an Array[StackTraceElement] for ease of logging.
+    */
+  private var stackAtFirstClose: Exception = null
+  private val stateChangeLock = new ReentrantReadWriteLock()
+
+  def get(key: Array[Byte]): Array[Byte] = withStateCheck {
     metrics.gets.inc
     require(key != null, "Null key not allowed.")
     val found = db.get(key)
@@ -113,7 +122,7 @@ class RocksDbKeyValueStore(
     found
   }
 
-  def getAll(keys: java.util.List[Array[Byte]]): java.util.Map[Array[Byte], Array[Byte]] = {
+  def getAll(keys: java.util.List[Array[Byte]]): java.util.Map[Array[Byte], Array[Byte]] = withStateCheck {
     metrics.getAlls.inc
     require(keys != null, "Null keys not allowed.")
     val map = db.multiGet(keys)
@@ -131,19 +140,19 @@ class RocksDbKeyValueStore(
     map
   }
 
-  def put(key: Array[Byte], value: Array[Byte]) {
+  def put(key: Array[Byte], value: Array[Byte]): Unit = withStateCheck {
     metrics.puts.inc
     require(key != null, "Null key not allowed.")
     if (value == null) {
       db.delete(writeOptions, key)
     } else {
-      metrics.bytesWritten.inc(key.size + value.size)
+      metrics.bytesWritten.inc(key.length + value.length)
       db.put(writeOptions, key, value)
     }
   }
 
   // Write batch from RocksDB API is not used currently because of: https://github.com/facebook/rocksdb/issues/262
-  def putAll(entries: java.util.List[Entry[Array[Byte], Array[Byte]]]) {
+  def putAll(entries: java.util.List[Entry[Array[Byte], Array[Byte]]]): Unit = withStateCheck {
     val iter = entries.iterator
     var wrote = 0
     var deletes = 0
@@ -164,25 +173,25 @@ class RocksDbKeyValueStore(
     metrics.deletes.inc(deletes)
   }
 
-  def delete(key: Array[Byte]) {
+  def delete(key: Array[Byte]): Unit = withStateCheck {
     metrics.deletes.inc
     put(key, null)
   }
 
-  def range(from: Array[Byte], to: Array[Byte]): KeyValueIterator[Array[Byte], Array[Byte]] = {
+  def range(from: Array[Byte], to: Array[Byte]): KeyValueIterator[Array[Byte], Array[Byte]] = withStateCheck {
     metrics.ranges.inc
     require(from != null && to != null, "Null bound not allowed.")
     new RocksDbRangeIterator(db.newIterator(), from, to)
   }
 
-  def all(): KeyValueIterator[Array[Byte], Array[Byte]] = {
+  def all(): KeyValueIterator[Array[Byte], Array[Byte]] = withStateCheck {
     metrics.alls.inc
     val iter = db.newIterator()
     iter.seekToFirst()
     new RocksDbIterator(iter)
   }
 
-  def flush {
+  def flush(): Unit = withStateCheck {
     metrics.flushes.inc
     trace("Flushing store: %s" format storeName)
     db.flush(flushOptions)
@@ -190,29 +199,53 @@ class RocksDbKeyValueStore(
   }
 
   def close() {
-    trace("Closing.")
-    db.close()
+    stateChangeLock.writeLock().lock()
+    try {
+      trace("Closing.")
+      if (stackAtFirstClose == null) { // first close
+        stackAtFirstClose = new Exception()
+        db.close()
+      } else {
+        warn(new SamzaException("Close called again on a closed store: %s. Ignoring this close." +
+          "Stack at first close is under 'Caused By'." format storeName, stackAtFirstClose))
+      }
+    } finally {
+      stateChangeLock.writeLock().unlock()
+    }
+  }
+
+  private def withStateCheck[T](fn: => T): T = {
+    stateChangeLock.readLock().lock()
+    try {
+      if (stackAtFirstClose == null) {
+        fn
+      } else {
+        throw new SamzaException("Attempted to access a closed store: %s. " +
+          "Stack at first close is under 'Caused By'." format storeName, stackAtFirstClose)
+      }
+    } finally {
+      stateChangeLock.readLock().unlock()
+    }
   }
 
   class RocksDbIterator(iter: RocksIterator) extends KeyValueIterator[Array[Byte], Array[Byte]] {
     private var open = true
-    private var firstValueAccessed = false
 
-    override def close() = {
+    override def close() = withStateCheck {
       open = false
       iter.close()
     }
 
     override def remove() = throw new UnsupportedOperationException("RocksDB iterator doesn't support remove")
 
-    override def hasNext() = iter.isValid
+    override def hasNext() = withStateCheck(iter.isValid)
 
     // The iterator is already pointing to the next element
-    protected def peekKey() = {
+    protected def peekKey() = withStateCheck {
       getEntry().getKey
     }
 
-    protected def getEntry() = {
+    protected def getEntry() = withStateCheck {
       val key = iter.key
       val value = iter.value
       new Entry(key, value)
@@ -224,21 +257,21 @@ class RocksDbKeyValueStore(
     // current element we are pointing to and advance the iterator to the next 
     // location (The new location may or may not be valid - this will surface
     // when the next next() call is made, the isValid will fail)
-    override def next() = {
+    override def next(): Entry[Array[Byte], Array[Byte]] = withStateCheck {
       if (!hasNext()) {
         throw new NoSuchElementException
       }
 
       val entry = getEntry()
-      iter.next
-      metrics.bytesRead.inc(entry.getKey.size)
+      iter.next()
+      metrics.bytesRead.inc(entry.getKey.length)
       if (entry.getValue != null) {
-        metrics.bytesRead.inc(entry.getValue.size)
+        metrics.bytesRead.inc(entry.getValue.length)
       }
       entry
     }
 
-    override def finalize() {
+    override def finalize(): Unit = withStateCheck {
       if (open) {
         trace("Leaked reference to RocksDB iterator, forcing close.")
         close()
@@ -250,9 +283,10 @@ class RocksDbKeyValueStore(
     // RocksDB's JNI interface does not expose getters/setters that allow the 
     // comparator to be pluggable, and the default is lexicographic, so it's
     // safe to just force lexicographic comparator here for now.
-    val comparator = lexicographic
-    iter.seek(from)
-    override def hasNext() = {
+    val comparator: LexicographicComparator = lexicographic
+    withStateCheck(iter.seek(from))
+
+    override def hasNext() = withStateCheck {
       super.hasNext() && comparator.compare(peekKey(), to) < 0
     }
   }