--- a/samza-test/src/main/config/perf/kv-perf.properties
+++ b/samza-test/src/main/config/perf/kv-perf.properties
@@ -37,6 +37,38 @@ test.rocksdb-write-performance.set-2.message.count=1000000
 test.rocksdb-write-performance.set-3.message.size=1024
 test.rocksdb-write-performance.set-3.message.count=1000000
 
+
+# Config for rocksdb-concurrent-write-performance
+test.rocksdb-concurrent-write-performance.stores.test-store.factory=org.apache.samza.storage.kv.RocksDbKeyValueStorageEngineFactory
+test.rocksdb-concurrent-write-performance.stores.test-store.object.cache.size=0
+test.rocksdb-concurrent-write-performance.partition.count=4
+
+test.rocksdb-concurrent-write-performance.set.count=6
+test.rocksdb-concurrent-write-performance.set-1.message.size=256
+test.rocksdb-concurrent-write-performance.set-1.message.count=1000000
+test.rocksdb-concurrent-write-performance.set-1.num.threads=4
+
+test.rocksdb-concurrent-write-performance.set-2.message.size=512
+test.rocksdb-concurrent-write-performance.set-2.message.count=1000000
+test.rocksdb-concurrent-write-performance.set-2.num.threads=4
+
+test.rocksdb-concurrent-write-performance.set-3.message.size=1024
+test.rocksdb-concurrent-write-performance.set-3.message.count=1000000
+test.rocksdb-concurrent-write-performance.set-3.num.threads=4
+
+test.rocksdb-concurrent-write-performance.set-4.message.size=256
+test.rocksdb-concurrent-write-performance.set-4.message.count=4000000
+test.rocksdb-concurrent-write-performance.set-4.num.threads=1
+
+test.rocksdb-concurrent-write-performance.set-5.message.size=512
+test.rocksdb-concurrent-write-performance.set-5.message.count=4000000
+test.rocksdb-concurrent-write-performance.set-5.num.threads=1
+
+test.rocksdb-concurrent-write-performance.set-6.message.size=1024
+test.rocksdb-concurrent-write-performance.set-6.message.count=4000000
+test.rocksdb-concurrent-write-performance.set-6.num.threads=1
+
+
 # Config for get-all-vs-get
 test.get-all-vs-get-write-many-read-many.stores.test-store.factory=org.apache.samza.storage.kv.RocksDbKeyValueStorageEngineFactory
 # Disable caching
@@ -52,4 +84,4 @@ test.get-all-vs-get-write-once-read-many.partition.count=4
 test.get-all-vs-get-write-once-read-many.set.count=3
 
 # List of tests to execute
-test.methods=rocksdb-write-performance,get-all-vs-get-write-many-read-many,get-all-vs-get-write-once-read-many
+test.methods=rocksdb-write-performance,rocksdb-concurrent-write-performance,get-all-vs-get-write-many-read-many,get-all-vs-get-write-once-read-many