Made RocksDB operations check if DB is still open to avoid segfaults.