@prateekm 2017-10-19
Some numbers from running the new concurrent write test on my MBP with and without these changes. These are not well controlled experiments, but I don't see any obvious perf degradations.

```
With Locks:
4 threads, 1000000 messages
[main] TestKeyValuePerformance$ [INFO] Running test: rocksdb-concurrent-write-performance
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 256 bytes with 4 threads: 4.861 sec
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 512 bytes with 4 threads: 9.721 sec
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 1024 bytes with 4 threads: 12.998000000000001 sec
[main] TestKeyValuePerformance [INFO] Total time to write 4000000 msgs of size 256 bytes with 1 threads: 8.55 sec
[main] TestKeyValuePerformance [INFO] Total time to write 4000000 msgs of size 512 bytes with 1 threads: 15.101 sec
[main] TestKeyValuePerformance [INFO] Total time to write 4000000 msgs of size 1024 bytes with 1 threads: 25.928 sec

8 threads, 8000000 messages
[main] TestKeyValuePerformance$ [INFO] Running test: rocksdb-concurrent-write-performance
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 256 bytes with 8 threads: 22.993000000000002 sec
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 512 bytes with 8 threads: 25.398 sec
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 1024 bytes with 8 threads: 41.051 sec
[main] TestKeyValuePerformance [INFO] Total time to write 8000000 msgs of size 256 bytes with 1 threads: 18.488 sec
[main] TestKeyValuePerformance [INFO] Total time to write 8000000 msgs of size 512 bytes with 1 threads: 31.264 sec
[main] TestKeyValuePerformance [INFO] Total time to write 8000000 msgs of size 1024 bytes with 1 threads: 64.231 sec

32 threads, 32000000 messages
[main] TestKeyValuePerformance$ [INFO] Running test: rocksdb-concurrent-write-performance
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 256 bytes with 32 threads: 119.179 sec
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 512 bytes with 32 threads: 140.784 sec
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 1024 bytes with 32 threads: 186.508 sec
[main] TestKeyValuePerformance [INFO] Total time to write 32000000 msgs of size 256 bytes with 1 threads: 125.158 sec
[main] TestKeyValuePerformance [INFO] Total time to write 32000000 msgs of size 512 bytes with 1 threads: 163.649 sec
[main] TestKeyValuePerformance [INFO] Total time to write 32000000 msgs of size 1024 bytes with 1 threads: 263.729 sec


Without Locks:
4 threads, 1000000 messages
[main] TestKeyValuePerformance$ [INFO] Running test: rocksdb-concurrent-write-performance
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 256 bytes with 4 threads: 5.125 sec
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 512 bytes with 4 threads: 10.417 sec
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 1024 bytes with 4 threads: 13.514000000000001 sec
[main] TestKeyValuePerformance [INFO] Total time to write 4000000 msgs of size 256 bytes with 1 threads: 8.374 sec
[main] TestKeyValuePerformance [INFO] Total time to write 4000000 msgs of size 512 bytes with 1 threads: 17.107 sec
[main] TestKeyValuePerformance [INFO] Total time to write 4000000 msgs of size 1024 bytes with 1 threads: 26.405 sec

8 threads, 8000000 messages
[main] TestKeyValuePerformance$ [INFO] Running test: rocksdb-concurrent-write-performance
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 256 bytes with 8 threads: 24.841 sec
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 512 bytes with 8 threads: 28.748 sec
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 1024 bytes with 8 threads: 39.853 sec
[main] TestKeyValuePerformance [INFO] Total time to write 8000000 msgs of size 256 bytes with 1 threads: 21.443 sec
[main] TestKeyValuePerformance [INFO] Total time to write 8000000 msgs of size 512 bytes with 1 threads: 32.207 sec
[main] TestKeyValuePerformance [INFO] Total time to write 8000000 msgs of size 1024 bytes with 1 threads: 54.419000000000004 sec

32 threads, 32000000 messages
[main] TestKeyValuePerformance$ [INFO] Running test: rocksdb-concurrent-write-performance
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 256 bytes with 32 threads: 108.47500000000001 sec
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 512 bytes with 32 threads: 129.094 sec
[main] TestKeyValuePerformance [INFO] Total time to write 1000000 msgs of size 1024 bytes with 32 threads: 189.90200000000002 sec
[main] TestKeyValuePerformance [INFO] Total time to write 32000000 msgs of size 256 bytes with 1 threads: 134.212 sec
[main] TestKeyValuePerformance [INFO] Total time to write 32000000 msgs of size 512 bytes with 1 threads: 164.27100000000002 sec
[main] TestKeyValuePerformance [INFO] Total time to write 32000000 msgs of size 1024 bytes with 1 threads: 290.944 sec
```