@prateekm 2017-10-25
@xinyuiscool It's not just closing the store twice, but any operation on the store after a close. I'll sync with you offline to understand the use case you mentioned.