@prateekm 2017-10-25
Talked to @xinyuiscool offline. The reader writer lock is for changing the state of the store (i.e. closing it), not for writing (put/flush) to it. get/put/flush/delete all acquire the read lock, so contention shouldn't be a huge issue.